app/startup-swipe-schedu/requirements.txt
snowflake-connector-python==4.0.0
